// src/lib/types/place.ts
export interface Place {
    contentid: number
    contenttypeid: number
    areacode: number
    sigungucode: number
    title: string
    addr1: string
    tel: string
    firstimage: string
    cat1: string
    cat2: string
    cat3: string
    lclsSystm1: string
    lclsSystm2: string
    lclsSystm3: string
    createdtime: string
    modifiedtime: string
    mapx: number
    mapy: number
    mlevel: number
}